vnfm-core: Move monitoring code.
